+++ src/controllers/list.controller.ts
@@ -5,19 +5,19 @@ import {
   FilterExcludingWhere,
 } from '@loopback/repository';
 import {
+  HttpErrors,
   del,
   get,
-  getModelSchemaRef,
   param,
   post,
-  requestBody,
-  response,
+  requestBody
 } from '@loopback/rest';
 import {SecurityBindings, UserProfile, securityId} from '@loopback/security';
 import {Auth} from '../auth/keys';
 import {List} from '../models';
 import {BaseEntryRepository, BaseListRepository} from '../repositories';
 import {Base} from '../repositories/keys';
+import {listSchema} from '../schemas';
 
 
 @injectable({tags: {name: 'ListController'}})
@@ -28,55 +28,40 @@ export class ListController {
     @inject(Base.Repository.ENTRY) private entryRepository: BaseEntryRepository
   ) { }
 
-  @post('/lists') //TODO introduce a cap on user requests though. In books and entries as well.
+  @post('/lists')
   async create(
+    @requestBody() request: any,
     @inject(SecurityBindings.USER)
     currentUserProfile: UserProfile,
-    @requestBody({
-      content: {
-        'application/json': {
-          schema: getModelSchemaRef(List, {
-            exclude: ['userId'],
-          }),
-        },
-      },
-    })
-    list: List,
   ): Promise<List> {
-    list.userId = currentUserProfile[securityId];
-    return this.listRepository.create(list);
+    const data = await listSchema.validateAsync(request);
+    data.userId = currentUserProfile[securityId];
+    return this.listRepository.create(data);
   }
 
   @authorize({voters: [Auth.Voter.OWN_LIST_ONLY]})
   @get('/lists/{listId}')
-  @response(200, {
-    description: 'List model instance',
-    content: {
-      'application/json': {
-        schema: getModelSchemaRef(List, { // TEST
-          exclude: ['userId'],
-        }),
-      },
-    },
-  })
   async findById(
     @param.path.number('listId') listId: number,
     @param.filter(List, {exclude: 'where'}) filter?: FilterExcludingWhere<List>
   ): Promise<List> {
+    const listExists = await this.entryRepository.exists(listId);
+    if (!listExists) {
+      throw new HttpErrors.NotFound('List not found');
+    }
     return this.listRepository.findById(listId, filter);
   }
 
   @authorize({voters: [Auth.Voter.OWN_LIST_ONLY]})
-  @del('/lists/{id}')
-  @response(204, {
-    description: 'List DELETE success',
-  })
+  @del('/lists/{listId}')
   async deleteById(
-    @param.path.number('id') listId: number
+    @param.path.number('listId') listId: number
   ): Promise<void> {
-    const filter = {where: {listId}}
-    await this.entryRepository.find(filter);
+    const listExists = await this.entryRepository.exists(listId);
+    if (!listExists) {
+      throw new HttpErrors.NotFound('List not found');
+    }
+    await this.entryRepository.deleteAll({listId});
     await this.listRepository.deleteById(listId);
-    // TODO Delete entries
   }
 }
